# GO-Manager/internal/http/docs/openapi.yaml
openapi: 3.0.0
info:
  title: GO-Manager API
  version: "1.0"
  description: API documentation for auth endpoints
servers:
  - url: http://localhost:8080
paths:
  /register:
    post:
      summary: Register user
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
              example:
                id: 123
                email: user@example.com
        '400':
          description: Bad request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_request
                  message: invalid request
                  details:
                    - field: email
                      code: required
                      message: email is required
  /login:
    post:
      summary: Login user
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                refresh_token: refresh-token-example
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_request
                  message: invalid request
                  details:
                    - field: password
                      code: required
                      message: password is required
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: unauthorized
                  message: invalid credentials
  /sessions:
    post:
      summary: Create session
      tags:
        - session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sessions/{id}/records:
    post:
      summary: Save record to session
      tags:
        - session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
      responses:
        '201':
          description: Record saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: record saved
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get session records
      tags:
        - session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Records list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    ErrorResponse:
      type: object
      example:
        error:
          code: invalid_request
          message: invalid request
          details:
            - field: email
              code: required
              message: email is required
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      properties:
        code:
          type: string
          example: invalid_request
        message:
          type: string
          example: invalid request
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: email
        code:
          type: string
          example: required
        message:
          type: string
          example: email is required
    CreateSessionRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    SessionResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    CreateRecordRequest:
      type: object
      properties:
        phrase:
          type: string
        context:
          type: string
      required:
        - phrase
        - context
    Context:
      type: object
      properties:
        phrase:
          type: string
        translation:
          type: string
    Record:
      type: object
      properties:
        phrase:
          type: string
        translations:
          type: array
          items:
            type: string
        synonyms:
          type: array
          items:
            type: string
        antonyms:
          type: array
          items:
            type: string
        baseForm:
          type: string
        contexts:
          type: array
          items:
            $ref: '#/components/schemas/Context'
    RecordsResponse:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/Record'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token as `Bearer <token>`